+++ lib/src/custom_packages/crop_image/crop_image.dart
@@ -1,7 +1,6 @@
 import 'dart:io';
 import 'dart:math';
 import 'dart:ui' as ui;
-import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'package:video_player/video_player.dart';
 
@@ -46,7 +45,7 @@ class CustomCrop extends StatefulWidget {
 }
 
 class CustomCropState extends State<CustomCrop>
-    with TickerProviderStateMixin, Drag {
+    with TickerProviderStateMixin {
   final _surfaceKey = GlobalKey();
 
   late final AnimationController _activeController;